# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Family Tree Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /families:
    post:
      operationId: createFamily
      tags: [families]
      summary: Create a new family tree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyInput"
      responses:
        "201":
          description: Family created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Family"
  /families/{familyId}:
    get:
      operationId: getFamily
      tags: [families]
      summary: Get family details
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Family details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Family"
    patch:
      operationId: updateFamily
      tags: [families]
      summary: Update family info
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Family"
  /families/{familyId}/stats:
    get:
      operationId: getFamilyStats
      tags: [families]
      summary: Get family statistics (member count, generations, etc.)
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Family stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyStats"
  /families/{familyId}/invite:
    post:
      operationId: generateInvite
      tags: [families]
      summary: Generate an invitation token (Gatekeeper only)
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "200":
          description: Invite token generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteToken"
  /families/{familyId}/pending-members:
    get:
      operationId: getPendingMembers
      tags: [families]
      summary: List pending approval members
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pending members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /families/{familyId}/pending-members/{userId}/approve:
    post:
      operationId: approveMember
      tags: [families]
      summary: Approve a pending member
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /families/{familyId}/pending-members/{userId}/reject:
    post:
      operationId: rejectMember
      tags: [families]
      summary: Reject a pending member
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member rejected

  # ── MEMBERS / TREE NODES ──────────────────────────────────────────────
  /families/{familyId}/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all members of a family tree
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Add a new member node to the tree
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
...
  /families/{familyId}/posts:
    get:
      operationId: getFamilyFeed
      tags: [posts]
      summary: Get the family social feed
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Posts feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostPage"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /families/{familyId}/media:
    get:
      operationId: getMediaGallery
      tags: [media]
      summary: Get family media gallery
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [image, video, audio]
      responses:
        "200":
          description: Media items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaItem"
    post:
      operationId: createMedia
      tags: [media]
      summary: Save a Cloudinary-uploaded media record
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaInput"
      responses:
        "201":
          description: Media saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
...
  /families/{familyId}/chronicle:
    post:
      operationId: generateChronicle
      tags: [families]
      summary: Generate downloadable PDF chronicle of family tree
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF generation job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChronicleJob"

  # ── ADMIN COCKPIT ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, displayName, status, role, familyId, createdAt]
      properties:
        id:          { type: string }
        email:       { type: string }
        displayName: { type: string }
        avatarUrl:   { type: ["string", "null"] }
        status:      { type: string, enum: [pending, active, suspended] }
        role:        { type: string, enum: [user, editor, gatekeeper, master_admin] }
        familyId:    { type: ["string", "null"] }
        memberId:    { type: ["string", "null"] }
        createdAt:   { type: string, format: date-time }

    # ── FAMILY SCHEMAS ──────────────────────────────────────────────────
    FamilyInput:
      type: object
      required: [name]
      properties:
        name:        { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        coverUrl:    { type: ["string", "null"] }
        theme:       { type: string, enum: [heritage, vibrant, cyber], default: heritage }
    FamilyUpdate:
      type: object
      properties:
        name:        { type: string }
        description: { type: ["string", "null"] }
        coverUrl:    { type: ["string", "null"] }
        theme:       { type: string, enum: [heritage, vibrant, cyber] }
    Family:
      type: object
      required: [id, name, familyTreeId, gatekeeperId, createdAt]
      properties:
        id:           { type: string }
        name:         { type: string }
        description:  { type: ["string", "null"] }
        coverUrl:     { type: ["string", "null"] }
        familyTreeId: { type: string }
        gatekeeperId: { type: string }
        theme:        { type: string, enum: [heritage, vibrant, cyber] }
        memberCount:  { type: integer }
        createdAt:    { type: string, format: date-time }
    FamilyStats:
      type: object
      required: [totalMembers, livingMembers, generations, totalPosts, totalMedia, birthdaysToday]
      properties:
        totalMembers:   { type: integer }
        livingMembers:  { type: integer }
        generations:    { type: integer }
        totalPosts:     { type: integer }
        totalMedia:     { type: integer }
        birthdaysToday: { type: array, items: { $ref: "#/components/schemas/Member" } }
    InviteInput:
      type: object
      properties:
        email:      { type: ["string", "null"] }
        expiresInH: { type: integer, default: 72 }
    InviteToken:
      type: object
      required: [token, url, expiresAt]
      properties:
        token:     { type: string }
        url:       { type: string }
        expiresAt: { type: string, format: date-time }

    # ── MEMBER SCHEMAS ──────────────────────────────────────────────────
    MemberInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:    { type: string }
        lastName:     { type: string }
        gender:       { type: ["string", "null"], enum: [male, female, other, null] }
        birthDate:    { type: ["string", "null"] }
        deathDate:    { type: ["string", "null"] }
        birthPlace:   { type: ["string", "null"] }
        bio:          { type: ["string", "null"] }
        avatarUrl:    { type: ["string", "null"] }
        posX:         { type: ["number", "null"] }
        posY:         { type: ["number", "null"] }
        linkedUserId: { type: ["string", "null"] }
...
    Member:
      type: object
      required: [id, familyId, firstName, lastName, createdAt]
      properties:
        id:           { type: string }
        familyId:     { type: string }
        firstName:    { type: string }
        lastName:     { type: string }
        gender:       { type: ["string", "null"] }
        birthDate:    { type: ["string", "null"] }
        deathDate:    { type: ["string", "null"] }
        birthPlace:   { type: ["string", "null"] }
        bio:          { type: ["string", "null"] }
        avatarUrl:    { type: ["string", "null"] }
        posX:         { type: ["number", "null"] }
        posY:         { type: ["number", "null"] }
        linkedUserId: { type: ["string", "null"] }
        role:         { type: ["string", "null"] }
        locationLat:  { type: ["number", "null"] }
        locationLng:  { type: ["number", "null"] }
        locationName: { type: ["string", "null"] }
        createdAt:    { type: string, format: date-time }
...
    PostInput:
      type: object
      required: [content]
      properties:
        content:       { type: string, minLength: 1 }
        mediaUrls:     { type: array, items: { type: string } }
        taggedMembers: { type: array, items: { type: string } }
        isCapsule:     { type: boolean, default: false }
        unlockAt:      { type: ["string", "null"], format: date-time }
    Post:
      type: object
      required: [id, familyId, authorId, content, reactions, commentCount, isCapsule, createdAt]
      properties:
        id:            { type: string }
        familyId:      { type: string }
        authorId:      { type: string }
        authorName:    { type: string }
        authorAvatar:  { type: ["string", "null"] }
        content:       { type: string }
        mediaUrls:     { type: array, items: { type: string } }
        taggedMembers: { type: array, items: { type: string } }
        reactions:     { $ref: "#/components/schemas/Reactions" }
        commentCount:  { type: integer }
        isCapsule:     { type: boolean }
        unlockAt:      { type: ["string", "null"] }
        isLocked:      { type: boolean }
        createdAt:     { type: string, format: date-time }
    PostPage:
      type: object
      required: [posts, hasMore]
      properties:
        posts:      { type: array, items: { $ref: "#/components/schemas/Post" } }
        nextCursor: { type: ["string", "null"] }
        hasMore:    { type: boolean }
    Reactions:
      type: object
      properties:
        like:      { type: integer }
        love:      { type: integer }
        celebrate: { type: integer }
...
    MediaInput:
      type: object
      required: [cloudinaryId, url, type]
      properties:
        cloudinaryId:  { type: string }
        url:           { type: string }
        thumbnailUrl:  { type: ["string", "null"] }
        type:          { type: string, enum: [image, video, audio] }
        title:         { type: ["string", "null"] }
        description:   { type: ["string", "null"] }
        taggedMembers: { type: array, items: { type: string } }
        duration:      { type: ["number", "null"] }
    MediaItem:
      type: object
      required: [id, familyId, uploaderId, cloudinaryId, url, type, createdAt]
      properties:
        id:            { type: string }
        familyId:      { type: string }
        uploaderId:    { type: string }
        uploaderName:  { type: string }
        cloudinaryId:  { type: string }
        url:           { type: string }
        thumbnailUrl:  { type: ["string", "null"] }
        type:          { type: string, enum: [image, video, audio] }
        title:         { type: ["string", "null"] }
        description:   { type: ["string", "null"] }
        taggedMembers: { type: array, items: { type: string } }
        duration:      { type: ["number", "null"] }
        createdAt:     { type: string, format: date-time }
...
    ChronicleJob:
      type: object
      required: [jobId, status]
      properties:
        jobId:  { type: string }
        status: { type: string, enum: [processing, ready, failed] }
        url:    { type: ["string", "null"] }