# lib/api-zod/src/generated/types/family.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Family Tree Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { FamilyTheme } from './familyTheme';

export interface Family {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  familyTreeId: string;
  gatekeeperId: string;
  theme?: FamilyTheme;
  memberCount?: number;
  createdAt: Date;
}